Add findPackageForDownload
Support only major versions

Fix typo in folder name
